refactor(ui): move toast listener into app components
